Moved post BL into manager and api models into api
